requirements.txt
asgiref==3.7.2
Django==5.0
django-taggit==5.0.1
Markdown==3.5.1
psycopg==3.1.18
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
